sizes

+++ 0x00-hello_world/6-size.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 /**
- * main - printf out sizes of data types in C
+ * main - print out sizes of data types in C
  * code by xomee
  * Return: 0
 */
@@ -12,10 +12,10 @@ int main(void)
 	long long int d;
 	float f;
 
-	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
-	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+	printf("size of a char: %zu byte(s)\n", sizeof(a));
+	printf("size of an int: %zu byte(s)\n", sizeof(b));
+	printf("size of a long int: %zu byte(s)\n", sizeof(c));
+	printf("size of a long long int: %zu byte(s)\n", sizeof(d));
+	printf("size of a float: %zu byte(s)\n", sizeof(f));
 	return (0);
 }
